requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cymem==2.0.7
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.5.0/es_core_news_sm-3.5.0-py3-none-any.whl
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
install==1.3.5
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.2
openai==0.26.5
packaging==23.0
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.5
pydub==0.25.1
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.28.2
rfc3986==1.5.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.9.0
srsly==2.4.6
thinc==8.1.8
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
yarl==1.8.2
